Feat: UI/UX improvements in NewUserCardsPage.tsx, overall better Enum implementation Fix: TransactionType updated from int to TransactionType Enum in Transaction.cs entity model Fix: Removed unnecessary castings after the model's Enum fix Add: Copy card number functionality in NewUserCardsPage Add: Change card name functionality in NewUserCardsPage Update: CardName column is now limited to 16 characters maximum (down form 20) Add: A stripe in cards' view with the corresponding color of the card's type

+++ TRKart.API/Controllers/SecureTransactionController.cs
@@ -514,7 +514,7 @@ namespace TRKart.API.Controllers
                 var transaction = await _context.Transaction
                     .Include(t => t.UserCard)
                     .FirstOrDefaultAsync(t => t.TransactionID == transactionId 
-                                            && t.TransactionType == (int)TransactionType.TopUp 
+                                            && t.TransactionType == TransactionType.TopUp 
                                             && t.UserCard.CustomerID == customerId.Value);
 
                 if (transaction == null)
